Do not check alignment of byte relocations
Since byte relocations are used for both loading the adresses of static data as well as loading the accumulator for indirect jumps, we cannot assume that the address which will be relocated necessarily will be 2-byte aligned (which is required for jal/branch). A check should be made for this in the future - possibly adding new relocations.